feat: add new `child-dependent` versionincrement field for helm resource
the intent of this change is to allow helm charts to have a version bump that
corresponds with the highest level of bump done by a dependency rather than
just a static major, minor, or patch increment. this will allow potential major
or breaking changes in chart dependencies from being masked by a lesser increment
done on the parent chart.