feat: add new `auto` versionincrement field for helmchart resource (#1593)
* feat: add new `child-dependent` versionincrement field for helm resource

the intent of this change is to allow helm charts to have a version bump that
corresponds with the highest level of bump done by a dependency rather than
just a static major, minor, or patch increment. this will allow potential major
or breaking changes in chart dependencies from being masked by a lesser increment
done on the parent chart.

* chore: change `child-dependent` to `auto`

update the `child-dependent` name to something a bit
more generic

* feat: change 'auto' behavior for non-semver inputs

Adapt the behavior of the `auto` version bump behavior to cause
a minor version bump if either old or new target version is not
a valid semver string. this will ensure that the legacy behavior
of the default minor version bump will continue.

* fix: support downgrade scenarios

in the `auto` version bump scenario we will now check
if the new semver major, minor, and patch versions are
not equal to the old ones rather than simply if the new
version is greater than the old one. this will address
the possibility of a downgrade scenario.

* feat: add catch-all for semver metadata adjustments

add a catch-all block for the `auto` version bump setting
that will do a minor release bump if the old and new semver
major, minor, and patch versions are identical which implies
that the change must be in the metadata tags.

---------

Co-authored-by: Olivier Vernin <olivier@vernin.me>