feat: Add support for GitHub API commit
This commit adds support for GitHub API commit.
It allows to update a file in a GitHub repository
using the GitHub API instead of using git.

The GitHub API commit is useful when a user
is using the automated token from Github Actions
as it creates a signed commit automatically.