feat(experimental): Expose given values to CUE as `#values`
Currently, CUE code can't interact with given values directly; instead,
the rendered yaml is then templated with the values, which isn't
particularly useful.

This instead will add all the values to the CUE code's scope as a new
definition named `#values`, which can then be accessed for the purpose
of generating the final manifest. Effectively, the "templating" becomes
the unification of the CUE manifest with values object.