Various feature additions for CUE (#5352)
* feat(experimental): Treat updatecli.cue as a default filename

We already have to pass --experimental, so also passing the filename
doesn't add very much.

* feat(experimental): Expose given values to CUE as `#values`

Currently, CUE code can't interact with given values directly; instead,
the rendered yaml is then templated with the values, which isn't
particularly useful.

This instead will add all the values to the CUE code's scope as a new
definition named `#values`, which can then be accessed for the purpose
of generating the final manifest. Effectively, the "templating" becomes
the unification of the CUE manifest with values object.

---------

Co-authored-by: Olivier Vernin <olivier@vernin.me>